Adding Favourite Payee functionality / Security Utils / Changing dashboard to use tabs

+++ src/main/java/com/bankingapp/backend/repository/CustomerRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long>, JpaSpecificationExecutor<Customer> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer>, JpaSpecificationExecutor<Customer> {
     Customer findByUsername(String username);
 }
